chore: split up folders
