+++ package.json
@@ -33,7 +33,6 @@
     "@trpc/react": "^9.17.0",
     "@trpc/server": "^9.17.0",
     "bcrypt": "^5.0.1",
-    "clsx": "^1.1.1",
     "next": "^12.0.1",
     "next-auth": "^4.1.0",
     "react": "^17.0.2",
