+++ src/lib/index.ts
@@ -1,0 +1,2 @@
+export { prisma } from './prisma'
+export * from './trpc'
