+++ src/lib/trpc.ts
@@ -1,0 +1,10 @@
+import { createReactQueryHooks } from '@trpc/react'
+import type{ AppRouter } from '@/server/routers/_app'
+
+import type { inferProcedureOutput } from '@trpc/server'
+
+export const trpc = createReactQueryHooks<AppRouter>()
+
+export type inferQueryOutput<TRouteKey extends keyof AppRouter['_def']['queries']> = inferProcedureOutput<
+    AppRouter['_def']['queries'][TRouteKey]
+>
