+++ src/server/lib/index.ts
@@ -1,0 +1,1 @@
+export { sum } from './sum'
