+++ src/server/routers/index.ts
@@ -1,0 +1,1 @@
+export * from './_app'
